Implementing CardController Attack and HP methods!
These are obviously subject to change if necessary, but these methods outline what exactly I'm trying to accomplish through abstraction of the BattleController superclass. For example, I can CardController.Attack() any BattleController, because it will have to implement (or throw and exception or something) some way to be attacked, if possible. The beauty of abstraction and inheritance!

:dromedary_camel:

+++ Assets/Klayble Unity Files/ScriptableObjects/AttackData.cs
@@ -9,7 +9,7 @@ public class AttackData : ScriptableObject
     public Elementals ElementalType { get => m_ElementalType; private set => m_ElementalType = value; }
 
     [SerializeField]
-    private float m_BaseDamage;
+    private int m_BaseDamage;
 
-    public float BaseDamage { get => m_BaseDamage; private set => m_BaseDamage = value; }
+    public int BaseDamage { get => m_BaseDamage; private set => m_BaseDamage = value; }
 }
